# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sharerr",
    "description": "The machine-facing HTTP API of a sharerr instance: the indexer feed a friend's Sonarr or Radarr queries, the BitTorrent tracker their client announces to, the gossip exchange that keeps addresses current, the lighthouse rendezvous, and the operational endpoints an orchestrator polls.\n\nThe server-rendered web UI is deliberately absent — see the `sharerr::openapi` module docs.\n\n**Authentication.** Every feed and gossip endpoint takes a peer's own API key as an `apikey` query parameter. Each friend holds a different one, so a single friend can be cut off without disturbing the others, and every request records who made it. There is no unauthenticated read path: a missing key is refused exactly the way a wrong one is, because saying \"this instance has no key configured\" would confirm the port belongs to sharerr.\n\n**Base URL.** None is listed. A sharerr instance is reached at whatever address its operator advertises — often behind a VPN with a rotating forwarded port, which is the whole reason the gossip and lighthouse endpoints exist.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/announce": {
      "get": {
        "tags": [
          "tracker"
        ],
        "summary": "`GET /announce`.",
        "operationId": "announce",
        "parameters": [
          {
            "name": "info_hash",
            "in": "query",
            "description": "The torrent's info hash: 20 raw bytes, percent-encoded. **Not hex.**",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "peer_id",
            "in": "query",
            "description": "The client's own id: 20 raw bytes, percent-encoded.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "port",
            "in": "query",
            "description": "The port the client accepts connections on.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 6881
          },
          {
            "name": "left",
            "in": "query",
            "description": "Bytes still to download. `0` means a seeder.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "`started`, `stopped`, `completed`, or absent for a periodic re-announce.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "compact",
            "in": "query",
            "description": "`1` for the compact peer list. Anything else returns the dictionary form.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "numwant",
            "in": "query",
            "description": "How many peers to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "ip",
            "in": "query",
            "description": "The client's own address. Honoured **only when it is a private address** —\notherwise the connection's source address wins, so a peer cannot announce\nsomebody else into the swarm.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A bencoded dictionary. **Always 200** — the BitTorrent tracker protocol reports refusal as a bencoded `failure reason` key, not as an HTTP status, so a client that only checks the status code will read an error as a swarm.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/announce/{token}": {
      "get": {
        "tags": [
          "tracker"
        ],
        "summary": "`GET /announce/{token}`.",
        "operationId": "announceWithToken",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "The announce token baked into this torrent's announce URL. Identifies which peer is announcing, so a swarm can be attributed and one friend's access revoked without disturbing anyone else's.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "info_hash",
            "in": "query",
            "description": "The torrent's info hash: 20 raw bytes, percent-encoded. **Not hex.**",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "peer_id",
            "in": "query",
            "description": "The client's own id: 20 raw bytes, percent-encoded.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "port",
            "in": "query",
            "description": "The port the client accepts connections on.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 6881
          },
          {
            "name": "left",
            "in": "query",
            "description": "Bytes still to download. `0` means a seeder.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "`started`, `stopped`, `completed`, or absent for a periodic re-announce.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "compact",
            "in": "query",
            "description": "`1` for the compact peer list. Anything else returns the dictionary form.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "numwant",
            "in": "query",
            "description": "How many peers to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "ip",
            "in": "query",
            "description": "The client's own address. Honoured **only when it is a private address** —\notherwise the connection's source address wins, so a peer cannot announce\nsomebody else into the swarm.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A bencoded dictionary. **Always 200** — the BitTorrent tracker protocol reports refusal as a bencoded `failure reason` key, not as an HTTP status, so a client that only checks the status code will read an error as a swarm.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api": {
      "get": {
        "tags": [
          "torznab"
        ],
        "summary": "`GET /api?t=...`",
        "operationId": "torznab",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "description": "The Torznab function. `caps` for the capabilities document, or one of the\nsearch functions: `search`, `tvsearch`, `movie`, `music`, `book`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tvsearch"
          },
          {
            "name": "q",
            "in": "query",
            "description": "Free-text needle, matched against the release title.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "ep",
            "in": "query",
            "description": "Daily shows send `ep=MM/DD` rather than a number. That form has no\nseason/episode to match against here, so it is read as \"no episode\nfilter\" instead of failing the whole request with a bare 400 — which\nProwlarr counts as an indexer failure and backs off from, hiding every\nother release too.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "tvdbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "tmdbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "imdbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Torznab document: the capabilities XML for `t=caps`, otherwise an RSS feed of matching releases. Each item carries a `.torrent` link and a magnet whose announce tiers are attributed to the calling peer.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "No such Torznab function — a Torznab `<error code=\"202\">`.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "No `apikey`, or one that matches no active peer. `t=caps` requires it too, deliberately: one fewer endpoint that says anything to an unauthenticated caller.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "The database is not open yet.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "peerApiKey": []
          }
        ]
      }
    },
    "/api/gossip/endpoints": {
      "get": {
        "tags": [
          "gossip"
        ],
        "summary": "`GET /api/gossip/endpoints?peers=pk1,pk2` — the pull side.",
        "operationId": "gossipPull",
        "parameters": [
          {
            "name": "peers",
            "in": "query",
            "description": "Comma-separated hex pubkeys the caller already knows, so the answer can skip them.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed endpoint records: this instance's own first, then any it holds for peers it shares with. Each is signed by the peer it describes, so nothing here has to be trusted on the relayer's word.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GossipRecordBatch"
                }
              }
            }
          },
          "401": {
            "description": "No `apikey`, or one that matches no active peer. Answered as Torznab's own XML error, the same as the feed — this rides the feed's authentication.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "The database is not open yet.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "peerApiKey": []
          }
        ]
      },
      "post": {
        "tags": [
          "gossip"
        ],
        "summary": "`POST /api/gossip/endpoints` — the push side, for a friend whose address\nchanged and who can therefore no longer be pulled from.",
        "operationId": "gossipPush",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GossipRecordBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "What the batch amounted to. Records about peers this instance does not share with are counted `unknown` and dropped by design, not rejected — a friend relaying their whole view is normal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GossipIngestSummary"
                }
              }
            }
          },
          "401": {
            "description": "No `apikey`, or one that matches no active peer.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "The database is not open yet.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "peerApiKey": []
          }
        ]
      }
    },
    "/api/v2.0/indexers": {
      "get": {
        "tags": [
          "jackett"
        ],
        "summary": "`GET /api/v2.0/indexers` — the list of indexers, which for sharerr is one.",
        "description": "Jackett supports `?configured=true`. sharerr's single indexer is always\nconfigured, so the filter is accepted and changes nothing — no `Query`\nextractor means every filter a client sends is accepted and ignored, because\na client that sends one and gets an error learns nothing useful.",
        "operationId": "jackettIndexers",
        "parameters": [
          {
            "name": "configured",
            "in": "query",
            "description": "Accepted and ignored. sharerr's single indexer is always configured, so every value of this filter yields the same one-element list.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A one-element list: sharerr itself.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JackettIndexer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No `apikey`, or one that matches no active peer.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "peerApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GossipRecordBatch": {
        "type": "object",
        "description": "The wire shape of both gossip endpoints' bodies.",
        "required": [
          "records"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GossipEndpointRecord"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "peerApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "The API key this instance issued to one friend. Issued from Settings → Peers, shown once, and stored only as a hash — losing one means issuing a new one, which is the correct behaviour for a bearer credential."
      }
    }
  },
...
}